requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
django-debug-toolbar==3.2.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
Pillow==9.0.0
six==1.16.0
sqlparse==0.4.2
